+++ outputs.tf
@@ -8,3 +8,8 @@ output "model" {
   description = "Full model."
   value       = local.model
 }
+
+output "sites" {
+  description = "List of sites to be managed"
+  value       = local.sites
+}